Protect entityMetadataMap from concurrent access #197 - code clean & decompose
